test(go): Split go datasource tests
