+++ lib/datasource/go/goproxy.spec.ts
@@ -3,6 +3,7 @@ import { loadFixture } from '../../../test/util';
 import * as memCache from '../../util/cache/memory';
 import {
   encodeCase,
+  getReleases,
   listVersions,
   parseGoproxy,
   parseNoproxy,
@@ -123,6 +124,154 @@ describe('datasource/go/goproxy', () => {
       expect(parseNoproxy('x/b').test('x/aba')).toBeFalse();
       expect(parseNoproxy('x/ab').test('x/aba')).toBeFalse();
       expect(parseNoproxy('x/ab[a-b]').test('x/aba')).toBeTrue();
+    });
+  });
+
+  describe('GOPROXY', () => {
+    const baseUrl = 'https://proxy.golang.org';
+
+    afterEach(() => {
+      delete process.env.GOPROXY;
+      delete process.env.GONOPROXY;
+      delete process.env.GOPRIVATE;
+    });
+
+    it('skips GONOPROXY and GOPRIVATE packages', async () => {
+      process.env.GOPROXY = baseUrl;
+      process.env.GOPRIVATE = 'github.com/google/*';
+
+      httpMock.scope('https://api.github.com/');
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toBeEmpty();
+      expect(res).toBeNull();
+    });
+
+    it('fetches release data from goproxy', async () => {
+      process.env.GOPROXY = baseUrl;
+
+      httpMock
+        .scope(`${baseUrl}/github.com/google/btree`)
+        .get('/@v/list')
+        .reply(200, 'v1.0.0\nv1.0.1\n')
+        .get('/@v/v1.0.0.info')
+        .reply(200, { Version: 'v1.0.0', Time: '2018-08-13T15:31:12Z' })
+        .get('/@v/v1.0.1.info')
+        .reply(200, { Version: 'v1.0.1', Time: '2019-10-16T16:15:28Z' });
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toMatchSnapshot();
+      expect(res?.releases).toMatchObject([
+        { releaseTimestamp: '2018-08-13T15:31:12Z', version: 'v1.0.0' },
+        { releaseTimestamp: '2019-10-16T16:15:28Z', version: 'v1.0.1' },
+      ]);
+    });
+
+    it('handles timestamp fetch errors', async () => {
+      process.env.GOPROXY = baseUrl;
+
+      httpMock
+        .scope(`${baseUrl}/github.com/google/btree`)
+        .get('/@v/list')
+        .reply(200, 'v1.0.0\nv1.0.1\n')
+        .get('/@v/v1.0.0.info')
+        .replyWithError('unknown')
+        .get('/@v/v1.0.1.info')
+        .reply(410);
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toMatchSnapshot();
+      expect(res?.releases).toMatchObject([
+        { version: 'v1.0.0' },
+        { version: 'v1.0.1' },
+      ]);
+    });
+
+    it('handles pipe fallback', async () => {
+      process.env.GOPROXY = `https://example.com|${baseUrl}`;
+
+      httpMock
+        .scope('https://example.com/github.com/google/btree')
+        .get('/@v/list')
+        .replyWithError('unknown');
+
+      httpMock
+        .scope(`${baseUrl}/github.com/google/btree`)
+        .get('/@v/list')
+        .reply(200, 'v1.0.0\nv1.0.1\n')
+        .get('/@v/v1.0.0.info')
+        .reply(200, { Version: 'v1.0.0', Time: '2018-08-13T15:31:12Z' })
+        .get('/@v/v1.0.1.info')
+        .reply(200, { Version: 'v1.0.1', Time: '2019-10-16T16:15:28Z' });
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toMatchSnapshot();
+      expect(res?.releases).toMatchObject([
+        { releaseTimestamp: '2018-08-13T15:31:12Z', version: 'v1.0.0' },
+        { releaseTimestamp: '2019-10-16T16:15:28Z', version: 'v1.0.1' },
+      ]);
+    });
+
+    it('handles comma fallback', async () => {
+      process.env.GOPROXY = [
+        'https://foo.example.com',
+        'https://bar.example.com',
+        baseUrl,
+      ].join(',');
+
+      httpMock
+        .scope('https://foo.example.com/github.com/google/btree')
+        .get('/@v/list')
+        .reply(404);
+
+      httpMock
+        .scope('https://bar.example.com/github.com/google/btree')
+        .get('/@v/list')
+        .reply(410);
+
+      httpMock
+        .scope(`${baseUrl}/github.com/google/btree`)
+        .get('/@v/list')
+        .reply(200, 'v1.0.0\nv1.0.1\n')
+        .get('/@v/v1.0.0.info')
+        .reply(200, { Version: 'v1.0.0', Time: '2018-08-13T15:31:12Z' })
+        .get('/@v/v1.0.1.info')
+        .reply(200, { Version: 'v1.0.1', Time: '2019-10-16T16:15:28Z' });
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toMatchSnapshot();
+      expect(res?.releases).toMatchObject([
+        { releaseTimestamp: '2018-08-13T15:31:12Z', version: 'v1.0.0' },
+        { releaseTimestamp: '2019-10-16T16:15:28Z', version: 'v1.0.1' },
+      ]);
+    });
+
+    it('short-circuits with comma fallback', async () => {
+      process.env.GOPROXY = [
+        'https://foo.example.com',
+        'https://bar.example.com',
+        'https://baz.example.com',
+        baseUrl,
+      ].join(',');
+
+      httpMock
+        .scope('https://foo.example.com/github.com/google/btree')
+        .get('/@v/list')
+        .reply(404);
+
+      httpMock
+        .scope('https://bar.example.com/github.com/google/btree')
+        .get('/@v/list')
+        .reply(410);
+
+      httpMock
+        .scope('https://baz.example.com/github.com/google/btree')
+        .get('/@v/list')
+        .replyWithError('unknown');
+
+      const res = await getReleases({ lookupName: 'github.com/google/btree' });
+      expect(httpMock.getTrace()).toMatchSnapshot();
+      expect(res).toBeNull();
     });
   });
 });
